2080_LLM/Psy_RAG_Agent/requirements.txt
torch>=2.0.0
transformers>=4.30.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
hydra-core>=1.3.0
omegaconf>=2.3.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2